Added encounter naming
Added the ability to name new encounters to the main menu

+++ TimeForBattle/TimeForBattle/ViewModel/CombatListViewModel.cs
@@ -6,13 +6,15 @@ namespace TimeForBattle.ViewModel;
 public partial class CombatListViewModel : BaseViewModel
 {
     public CreatureService<Combat> CombatService;
+    public DialogService DialogService;
     public ObservableCollection<Combat> Combats { get; }
     [ObservableProperty] bool isRefreshing;
 
-    public CombatListViewModel(CreatureService<Combat> combatService)
+    public CombatListViewModel(CreatureService<Combat> combatService, DialogService dialogService)
     {
         Title = "Load Combat";
         this.CombatService = combatService;
+        this.DialogService = dialogService;
         Combats = [];
     }
 
@@ -34,7 +36,9 @@ public partial class CombatListViewModel : BaseViewModel
         if (combat is null)
             return;
 
-        await CombatService.DeleteAsync(combat);
+        bool answer = await DialogService.ShowConfirmationAsync((ContentPage)AppShell.Current.CurrentPage, "Delete?", "Are you sure you want to delete this encounter?", "Yes", "No");
+        if (answer)
+            await CombatService.DeleteAsync(combat);
 
         //also delete creatures in initiative for that combat?
